pdf-data-extractor/requirements.txt
openai>=1.51.0
python-dotenv>=1.0.1
pandas>=2.2.0
PyPDF2>=3.0.1
pdfplumber>=0.11.0
pdf2image>=1.17.0